requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
gradio_client==1.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
packaging==24.1
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
telebot==0.0.5
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
websockets==11.0.3
